+++ changelog.md
@@ -1,4 +1,4 @@
 Version: 23-04-2024_1 Alpha
 Changes:
 - Physics2D rework.
-    - Capsule collider added
+    - Fixed bug in handling b2Contact events